Move ObjectPart to the object module
