+++ mountpoint-s3/src/prefetch/caching_stream.rs
@@ -9,8 +9,7 @@ use tracing::{debug_span, trace, warn, Instrument};
 
 use crate::checksums::ChecksummedBytes;
 use crate::data_cache::{BlockIndex, DataCache};
-use crate::object::ObjectId;
-use crate::prefetch::part::ObjectPart;
+use crate::object::{ObjectId, ObjectPart};
 use crate::prefetch::part_queue::{unbounded_part_queue, PartQueueProducer};
 use crate::prefetch::part_stream::{ObjectPartStream, RequestRange};
 use crate::prefetch::task::RequestTask;
